[FEAT] 데이터베이스 상태 카드 모니터링 강화
- Uptime, 연결 수, QPS 등 실시간 메트릭 추가
- CPU, HEAP, Network 사용량 모니터링
- Heartbeat 시간 추적

+++ src/main/java/org/ravo/ravomanager/manager/dto/DatabaseStatusDto.java
@@ -14,7 +14,7 @@ public class DatabaseStatusDto {
     private String status;            // "Active", "Standby", "Down"
     private String uptime;            // "1d 0h 0m"
     private Integer connections;      // 현재 연결 수
-    private Integer qps;              // Queries per second
+    private Double qps;               // Queries per second (소수점 포함)
     private String latency;           // "2.3ms"
     private String lastHeartbeat;     // 마지막 하트비트 시간
     private Boolean isHealthy;        // 헬스 체크 상태
